+++ vns-cpp/compiler.hpp
@@ -6,12 +6,12 @@
 #include "libs/nlohmann/json.hpp"
 #include "scriptProcessor.hpp"
 
-using DialogueFileDataType = std::unordered_map<std::string, std::variant<DialogueSectionsDataType, std::unordered_map<std::string, int>, std::string>>;
+using DialogueFileDataType = std::unordered_map<std::string, std::variant<DialogueSectionsDataType, std::unordered_map<std::string, size_t>, std::string>>;
 
 class Compiler
 {
 public:
-    static std::unordered_map<std::string, int> get_compiler_info();
+    static std::unordered_map<std::string, size_t> get_compiler_info();
 
     static DialogueFileDataType load(const std::filesystem::path &);
