+++ vns-cpp/vns.cpp
@@ -13,7 +13,7 @@ int main(const int argc, char **argv)
     const std::unordered_set<std::string> arguments_with_input = {"-i", "-o", "-n"};
     const std::unordered_set<std::string> arguments_without_input = {"-h", "-s", "-t"};
     std::unordered_map<std::string, std::string> arguments_map;
-    for (int i = 1; i < argc; i++)
+    for (size_t i = 1; i < argc; i++)
     {
         if (std::string current_arg = argv[i]; arguments_with_input.contains(current_arg))
         {
